WIP: add benchmark_score.py file
